requirements.txt
Flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
Flask-Bcrypt  
Flask-JWT-Extended 